# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Firebox VCF Creator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CampaignStatus = typeof CampaignStatus[keyof typeof CampaignStatus];


export const CampaignStatus = {
  draft: 'draft',
  active: 'active',
  completed: 'completed',
} as const;

export interface Campaign {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: CampaignStatus;
  targetContacts: number;
  contactsCollected: number;
  remainingContacts?: number;
  progressPercent?: number;
  shareToken: string;
  shareLink: string;
  vcfDownloaded?: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CampaignInputStatus = typeof CampaignInputStatus[keyof typeof CampaignInputStatus];


export const CampaignInputStatus = {
  draft: 'draft',
  active: 'active',
  completed: 'completed',
} as const;

export interface CampaignInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  /** @minimum 1 */
  targetContacts: number;
  status?: CampaignInputStatus;
}

export type CampaignUpdateStatus = typeof CampaignUpdateStatus[keyof typeof CampaignUpdateStatus];


export const CampaignUpdateStatus = {
  draft: 'draft',
  active: 'active',
  completed: 'completed',
} as const;

export interface CampaignUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  /** @minimum 1 */
  targetContacts?: number;
  status?: CampaignUpdateStatus;
}

export interface Contact {
  id: number;
  campaignId: number;
  name: string;
  phone: string;
  submittedAt: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 7 */
  phone: string;
  consent: boolean;
}

export interface DashboardStats {
  totalCampaigns: number;
  activeCampaigns: number;
  completedCampaigns: number;
  draftCampaigns: number;
  totalContactsCollected: number;
  downloadedVcfs: number;
}

export interface CampaignAnalytics {
  id: number;
  title: string;
  status: string;
  targetContacts: number;
  contactsCollected: number;
  progressPercent: number;
  createdAt: string;
}