# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Firebox VCF Creator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: campaigns
    description: VCF Campaign management
  - name: contacts
    description: Contact submission for campaigns
  - name: dashboard
    description: Dashboard statistics and analytics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "400":
          description: Validation error

  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a single campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Campaign updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /campaigns/{id}/vcf:
    get:
      operationId: downloadCampaignVcf
      tags: [campaigns]
      summary: Download VCF file for a campaign (only when target reached)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: VCF file content
          content:
            text/vcard:
              schema:
                type: string
        "403":
          description: Target not yet reached
        "404":
          description: Not found

  /campaigns/share/{token}:
    get:
      operationId: getCampaignByToken
      tags: [campaigns]
      summary: Get public campaign info by share token (for submission page)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign public info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignPublic"
        "404":
          description: Not found

  /campaigns/{id}/contacts:
    get:
      operationId: listCampaignContacts
      tags: [contacts]
      summary: List contacts submitted to a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: submitContact
      tags: [contacts]
      summary: Submit a contact to a campaign (public endpoint)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Contact submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "409":
          description: Duplicate phone number
        "400":
          description: Validation error

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get overall dashboard statistics
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /analytics/campaigns:
    get:
      operationId: getCampaignAnalytics
      tags: [dashboard]
      summary: Get per-campaign analytics data
      responses:
        "200":
          description: Analytics data per campaign
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignAnalytics"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Campaign:
      type: object
      required:
        - id
        - title
        - status
        - targetContacts
        - contactsCollected
        - shareToken
        - shareLink
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, active, completed]
        targetContacts:
          type: integer
        contactsCollected:
          type: integer
        remainingContacts:
          type: integer
        progressPercent:
          type: number
        shareToken:
          type: string
        shareLink:
          type: string
        vcfDownloaded:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CampaignInput:
      type: object
      required:
        - title
        - targetContacts
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        targetContacts:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [draft, active, completed]

    CampaignUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        targetContacts:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [draft, active, completed]

    Contact:
      type: object
      required:
        - id
        - campaignId
        - name
        - phone
        - submittedAt
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        name:
          type: string
        phone:
          type: string
        submittedAt:
          type: string
          format: date-time

    ContactInput:
      type: object
      required:
        - name
        - phone
        - consent
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 7
        consent:
          type: boolean

    DashboardStats:
      type: object
      required:
        - totalCampaigns
        - activeCampaigns
        - totalContactsCollected
        - downloadedVcfs
        - completedCampaigns
        - draftCampaigns
      properties:
        totalCampaigns:
          type: integer
        activeCampaigns:
          type: integer
        completedCampaigns:
          type: integer
        draftCampaigns:
          type: integer
        totalContactsCollected:
          type: integer
        downloadedVcfs:
          type: integer

    CampaignPublic:
      type: object
      required:
        - id
        - title
        - status
        - targetContacts
        - contactsCollected
        - remainingContacts
        - progressPercent
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, active, completed]
        targetContacts:
          type: integer
        contactsCollected:
          type: integer
        remainingContacts:
          type: integer
        progressPercent:
          type: number

    CampaignAnalytics:
      type: object
      required:
        - id
        - title
        - status
        - targetContacts
        - contactsCollected
        - progressPercent
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        status:
          type: string
        targetContacts:
          type: integer
        contactsCollected:
          type: integer
        progressPercent:
          type: number
        createdAt:
          type: string
          format: date-time