# lib/api-zod/src/generated/types/campaign.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Firebox VCF Creator API
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignStatus } from './campaignStatus';

export interface Campaign {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: CampaignStatus;
  targetContacts: number;
  contactsCollected: number;
  remainingContacts?: number;
  progressPercent?: number;
  shareToken: string;
  shareLink: string;
  vcfDownloaded?: boolean;
  createdAt: Date;
  updatedAt: Date;
}