# lib/api-zod/src/generated/types/campaignAnalytics.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Firebox VCF Creator API
 * OpenAPI spec version: 0.1.0
 */

export interface CampaignAnalytics {
  id: number;
  title: string;
  status: string;
  targetContacts: number;
  contactsCollected: number;
  progressPercent: number;
  createdAt: Date;
}