# lib/api-zod/src/generated/types/campaignInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Firebox VCF Creator API
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignInputStatus } from './campaignInputStatus';

export interface CampaignInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  /** @minimum 1 */
  targetContacts: number;
  status?: CampaignInputStatus;
}