# lib/api-zod/src/generated/types/campaignUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Firebox VCF Creator API
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignUpdateStatus } from './campaignUpdateStatus';

export interface CampaignUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  /** @minimum 1 */
  targetContacts?: number;
  status?: CampaignUpdateStatus;
}