# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Firebox VCF Creator API
 * OpenAPI spec version: 0.1.0
 */

export interface Contact {
  id: number;
  campaignId: number;
  name: string;
  phone: string;
  submittedAt: Date;
}