# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Firebox VCF Creator API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 7 */
  phone: string;
  consent: boolean;
}